app/requirements.txt
django==3.2
python-dotenv==0.20.0
psycopg2-binary==2.9
django-split-settings==1.1.0
django-debug-toolbar==3.4.0
django-cors-headers==3.12.0
uwsgi==2.0.20
